Move WTPage to btree_format
