+++ tools/py_common/__init__.py
@@ -1,5 +1,5 @@
 # py_common/__init__.py
-from py_common.decoder import WTPage
+from py_common.btree_format import WTPage
 from py_common.stats import PageStats
 from py_common.printer import Printer
 from py_common.input import encode_bytes
